+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Removed
+- Remove DataTpl::lastChild field and associated methods
+
 ## [3.2.0] - 2024-08-27
 
 ### Fixed